third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGPatternElement-svgdom-x-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' property of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.x.baseVal.value is 50
PASS patternElement.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
